Add complete intelligence layer for trading signal generation
Integrates SMC pattern detection, feature engineering, and ML prediction into a unified pipeline.

+++ src/core/smc_engine.py
@@ -53,7 +53,7 @@ class SMCEngine:
             )
         
         # ATR is EMA of TR
-        atr = np.mean(tr[-period:])
+        atr = float(np.mean(tr[-period:]))
         return max(atr, 0.001)  # Avoid zero division
     
     @staticmethod
